pyproject.toml
[project]
name = "pde_module"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.6",
    "numpy>=2.3.2",
    "warp-lang>=1.9.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]

[tool.uv.workspace]
members = [
    ".",
]
[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"